requirements.txt
# --- Computer Vision & Hand Tracking ---
opencv-python
mediapipe
cvzone

# --- Data Handling & Math ---
numpy
pandas

# --- Machine Learning (Classifier) ---
tensorflow
scikit-learn

# --- Audio & Speech Recognition ---
openai-whisper
torch
torchaudio
sounddevice